Add unified streaming runner with resilient reconnects
